+++ src/disk-station/video-station/play-with-vlc/main.ts
@@ -1,4 +1,4 @@
-import { isVideoStationPage } from '../../lib'
+import { isVideoStationPage } from '../../util'
 import { handleVideoStation } from './handleVideoStation'
 
 if (isVideoStationPage(new URL(location.href))) {
